requirements.txt
pandas
matplotlib
fpdf
google-generativeai
python-dotenv
reportlab
beautifulsoup4
